trying to modularize commonly used functions
