+++ src/pm_struct.h
@@ -24,6 +24,20 @@ struct ADC_DATA {
   double  Volts    = 0.0;               // formatted value
 };
 
+union ulongArray
+{
+    uint32_t longNumber=0;
+    uint8_t  byteArray[4];
+};
+
+union floatArray
+{
+    float   floatNumber=0.0;
+    uint8_t byteArray[4];
+};
+
+
+
 volatile I2C_RX_DATA rxData;
 volatile I2C_TX_DATA txData;
 volatile ADC_DATA    adcDataBuffer[adcBufferSize];  // Enough room to store three adc readings
